+++ contracts/actions/furucombov2/IFurucomboV2.sol
@@ -1,0 +1,12 @@
+// SPDX-License-Identifier: MIT
+pragma solidity ^0.6.0;
+pragma experimental ABIEncoderV2;
+
+interface IFurucomboV2 {
+    function batchExec(
+        address[] calldata tos,
+        bytes32[] calldata configs,
+        bytes[] memory datas,
+        uint256[] calldata ruleIndexes
+    ) external payable;
+}